Adjusting the docking branch to generate internal cimgui bindings
For cimgui to generate the docking branch currently we need to pull some struct definitions into imgui_internal.h

When pulling ImGuiDockPreviewData, it's discovered the current cimgui generator doesn't peak at imgui.h to collect its globals, so when trying to parse ImGuiDockPreviewData, it chokes when reading the DropRects static array size expression that contains an enum from imgui.h (ImGuiDir_COUNT), so this has been hardcoded. It will need to be updated to be correct when fastforwarding, or this needs to be fixed in the cimgui generator.